# Ride-Now/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RideNow API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rides:
    get:
      operationId: getRides
      tags: [rides]
      summary: List rides (filtered by role)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: customerId
          in: query
          schema:
            type: ["integer", "null"]
        - name: driverId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of rides
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ride"

    post:
      operationId: createRide
      tags: [rides]
      summary: Request a new ride
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRideBody"
      responses:
        "201":
          description: Ride created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ride"
  /rides/{id}:
    get:
      operationId: getRide
      tags: [rides]
      summary: Get a ride by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ride details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ride"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rides/{id}/status:
    patch:
      operationId: updateRideStatus
      tags: [rides]
      summary: Update ride status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRideStatusBody"
      responses:
        "200":
          description: Updated ride
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ride"
  /rides/{id}/cancel:
    patch:
      operationId: cancelRide
      tags: [rides]
      summary: Cancel a ride
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled ride
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ride"
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /vehicles:
    get:
      operationId: getVehicles
      tags: [vehicles]
      summary: Get all vehicles (admin)
      responses:
        "200":
          description: Vehicles list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vehicle"

    post:
      operationId: createVehicle
      tags: [vehicles]
      summary: Create a vehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVehicleBody"
      responses:
        "201":
          description: Vehicle created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
  /vehicles/{id}:
    patch:
      operationId: updateVehicle
      tags: [vehicles]
      summary: Update vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVehicleBody"
      responses:
        "200":
          description: Updated vehicle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [customer, driver, admin]
        phone:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - createdAt
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
    Ride:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        driverId:
          type: ["integer", "null"]
        pickupAddress:
          type: string
        dropoffAddress:
          type: string
        rideType:
          type: string
          enum: [cab, bike]
        status:
          type: string
          enum: [requested, accepted, ongoing, completed, cancelled]
        fare:
          type: ["number", "null"]
        distance:
          type: ["number", "null"]
        customerName:
          type: ["string", "null"]
        driverName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - customerId
        - pickupAddress
        - dropoffAddress
        - rideType
        - status
        - createdAt
        - updatedAt
    CreateRideBody:
      type: object
      properties:
        pickupAddress:
          type: string
        dropoffAddress:
          type: string
        rideType:
          type: string
          enum: [cab, bike]
        distance:
          type: number
      required:
        - pickupAddress
        - dropoffAddress
        - rideType
        - distance
    UpdateRideStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [accepted, ongoing, completed, cancelled]
      required:
        - status
...
    Vehicle:
      type: object
      properties:
        id:
          type: integer
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        licensePlate:
          type: string
        vehicleType:
          type: string
          enum: [cab, bike, suv, van]
        isAvailable:
          type: boolean
        ratePerHour:
          type: ["number", "null"]
        ratePerKm:
          type: ["number", "null"]
        driverId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - make
        - model
        - year
        - licensePlate
        - vehicleType
        - isAvailable
        - createdAt
    CreateVehicleBody:
      type: object
      properties:
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        licensePlate:
          type: string
        vehicleType:
          type: string
          enum: [cab, bike, suv, van]
        ratePerHour:
          type: ["number", "null"]
        ratePerKm:
          type: ["number", "null"]
        driverId:
          type: ["integer", "null"]
      required:
        - make
        - model
        - year
        - licensePlate
        - vehicleType
    UpdateVehicleBody:
      type: object
      properties:
        isAvailable:
          type: boolean
        ratePerHour:
          type: ["number", "null"]
        ratePerKm:
          type: ["number", "null"]
        driverId:
          type: ["integer", "null"]
...